requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
feedparser==6.0.10

# AI Integration
openai==1.3.7
anthropic==0.7.7
langchain==0.0.350
transformers==4.36.0

# NLP and Text Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
newspaper3k==0.2.8

# SEO and Content
python-slugify==8.0.1
metapy==0.2.15

# Async Processing
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Data Processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Configuration
python-decouple==3.8
pyyaml==6.0.1

# Monitoring and Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2